refactor(profile): remove compatibility wrappers and use barrel imports
