+++ src/screens/profile/edit/index.ts
@@ -1,0 +1,1 @@
+export { ProfileEditScreen } from './view/ProfileEditScreen';
