+++ src/screens/profile/home/index.ts
@@ -1,0 +1,1 @@
+export { ProfileScreen } from './view/ProfileScreen';
