+++ src/screens/profile/index.ts
@@ -1,0 +1,5 @@
+export { ProfileEditScreen } from './edit';
+export { ProfileScreen } from './home';
+export { SearchProfilesScreen } from './search';
+export { UserProfileScreen } from './user';
+export * from './routes';
