+++ src/screens/profile/search/index.ts
@@ -1,0 +1,1 @@
+export { SearchProfilesScreen } from './view/SearchProfilesScreen';
