+++ src/screens/profile/user/index.ts
@@ -1,0 +1,1 @@
+export { UserProfileScreen } from './view/UserProfileScreen';
